Add files via upload
This is my Data Science project. I have worked on one of the project from Kaggle.com. I have worked on the classification algorithm which is Logistic Regression for identifying the tested patients whether the diabetics is positive or negative. 
Looking for the feedback on this project and the guidelines for improvements.